Updated data_entry::from_c_data() function to use u16 instead of i32

+++ stat_diary_tools/src/c_wrapper.rs
@@ -397,8 +397,8 @@ pub unsafe extern "C" fn InsertDataEntry(
     year: i32,
     month: i32,
     day: i32,
-    data: *const i32,
-    data_length: i32,
+    data: *const u16,
+    data_length: u32,
 ) -> i32 {
     if data.is_null() {
         return -3;
@@ -456,8 +456,8 @@ pub unsafe extern "C" fn AddDataEntry(
     year: i32,
     month: i32,
     day: i32,
-    data: *const i32,
-    data_length: i32,
+    data: *const u16,
+    data_length: u32,
 ) -> i32 {
     if data.is_null() {
         return -3;
